refactor document loaders to global util
